[added] enforceFocus prop to Modal
Allows you to configure whether the modal should enforce focus when
open. Ideally this doesn't need to be a public API, but in the case of
static models (like in the docs) you need to turn it off, because you
can't assume its the only one open on the page.